Refactor extra API calls and default Cartesia-Version to 2026-03-01.
Replace rest_client with extra_api helpers on the SDK httpx client so auth
and version stay consistent. Patch Fern client headers via sdk_setup.
